Rename some test files
